implement simpler record captive portal activity
