+++ app/src/main/res/layout/activity_record_captive_portal.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     style="@style/CaptivePortalAutoLogin.Edge2EdgeFix"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:animateLayoutChanges="true"
     android:orientation="vertical"
     >
     
@@ -13,7 +11,5 @@
         android:id="@+id/geckoView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:visibility="invisible"
-        tools:visibility="visible"
         />
 </LinearLayout>
